Add files via upload
All the file in same order with so you can refer them if you are newbie and just learned python and wanted top learn and explore about data base. I would also recommend you to download sqlite browser to view your database and look at the changes that you are making in the database.